Lenet FPGA example
